Fix - sync yaml without need to open code editor
